Add damage source selection UI
