+++ src/sidebar/buildings/BuildingsSection.tsx
@@ -1,12 +1,12 @@
-import { Box } from '@mui/material';
 import { StyleSelection } from 'sidebar/StyleSelection';
+import { SidebarPanelSection } from 'sidebar/ui/SidebarPanelSection';
 import { buildingsStyleState } from 'state/buildings';
 import { SidebarPanel } from '../SidebarPanel';
 
 export const BuildingsSection = () => {
   return (
     <SidebarPanel id="buildings" title="Buildings">
-      <Box p={2} bgcolor="#eee">
+      <SidebarPanelSection variant="style">
         <StyleSelection
           state={buildingsStyleState}
           options={[
@@ -17,7 +17,7 @@ export const BuildingsSection = () => {
           ]}
           defaultValue="type"
         />
-      </Box>
+      </SidebarPanelSection>
     </SidebarPanel>
   );
 };
