+++ src/sidebar/regions/RegionsSection.tsx
@@ -1,20 +1,20 @@
 import { FC } from 'react';
 import { RegionsControl } from './RegionsControl';
 import { SidebarPanel } from 'sidebar/SidebarPanel';
-import { Box } from '@mui/material';
 import { StyleSelection } from 'sidebar/StyleSelection';
 import { regionsStyleState } from 'state/regions';
 import { REGION_DEFAULT_STYLE, REGION_STYLES } from 'config/regions/styles';
+import { SidebarPanelSection } from 'sidebar/ui/SidebarPanelSection';
 
 export const RegionsSection: FC<{}> = () => {
   return (
     <SidebarPanel id="regions" title="Regions">
-      <Box p={2}>
+      <SidebarPanelSection>
         <RegionsControl />
-      </Box>
-      <Box p={2} bgcolor="#eee">
+      </SidebarPanelSection>
+      <SidebarPanelSection variant="style">
         <StyleSelection state={regionsStyleState} options={REGION_STYLES} defaultValue={REGION_DEFAULT_STYLE} />
-      </Box>
+      </SidebarPanelSection>
     </SidebarPanel>
   );
 };
