+++ src/sidebar/ui/InputSection.tsx
@@ -1,0 +1,7 @@
+import { Box } from '@mui/material';
+
+export const InputSection = ({ children }) => (
+  <Box mb={2} flexGrow={1} width="100%">
+    {children}
+  </Box>
+);
