+++ src/sidebar/ui/params/RCPControl.tsx
@@ -1,0 +1,25 @@
+import { FormControl, FormLabel, MenuItem, Select } from '@mui/material';
+import { DataParam } from './DataParam';
+
+function rcpLabel(value) {
+  return value === 'baseline' ? 'Baseline' : value;
+}
+
+export const RCPControl = ({ group }) => {
+  return (
+    <FormControl fullWidth>
+      <FormLabel>RCP</FormLabel>
+      <DataParam group={group} id="rcp">
+        {({ value, onChange, options }) => (
+          <Select variant="standard" value={value} onChange={(e) => onChange(e.target.value)} fullWidth>
+            {options.map((rcp) => (
+              <MenuItem key={rcp} value={rcp}>
+                {rcpLabel(rcp)}
+              </MenuItem>
+            ))}
+          </Select>
+        )}
+      </DataParam>
+    </FormControl>
+  );
+};
